+++ app/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@moondreamsdev/dreamer-ui": "^1.7.7-test.14",
+    "@moondreamsdev/dreamer-ui": "^1.7.7-test.15",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
