--- app/src/select/icons.tsx
@@ -1,58 +1,0 @@
-// Simple icons for the Select component
-import React from 'react';
-
-interface IconProps {
-  size?: number;
-  className?: string;
-}
-
-export const ChevronDown: React.FC<IconProps> = ({ size = 16, className = '' }) => (
-  <svg
-    width={size}
-    height={size}
-    className={className}
-    viewBox='0 0 15 15'
-    xmlns='http://www.w3.org/2000/svg'
-  >
-    <path
-      d='M3.13523 6.15803C3.3241 5.95657 3.64052 5.94637 3.84197 6.13523L7.5 9.56464L11.158 6.13523C11.3595 5.94637 11.6759 5.95657 11.8648 6.15803C12.0536 6.35949 12.0434 6.67591 11.842 6.86477L7.84197 10.6148C7.64964 10.7951 7.35036 10.7951 7.15803 10.6148L3.15803 6.86477C2.95657 6.67591 2.94637 6.35949 3.13523 6.15803Z'
-      fill='currentColor'
-      fillRule='evenodd'
-      clipRule='evenodd'
-    />
-  </svg>
-);
-
-export const Check: React.FC<IconProps> = ({ size = 16, className = '' }) => (
-  <svg
-    width={size}
-    height={size}
-    className={className}
-    viewBox='0 0 15 15'
-    xmlns='http://www.w3.org/2000/svg'
-  >
-    <path
-      d='M11.4669 3.72684C11.7558 3.91574 11.8369 4.30308 11.648 4.59198L7.39799 11.092C7.29783 11.2452 7.13556 11.3467 6.95402 11.3699C6.77247 11.3931 6.58989 11.3355 6.45446 11.2124L3.70446 8.71241C3.44905 8.48022 3.43023 8.08494 3.66242 7.82953C3.89461 7.57412 4.28989 7.55529 4.5453 7.78749L6.75292 9.79441L10.6018 3.90792C10.7907 3.61902 11.178 3.53795 11.4669 3.72684Z'
-      fill='currentColor'
-      fillRule='evenodd'
-      clipRule='evenodd'
-    />
-  </svg>
-);
-
-export const X: React.FC<IconProps> = ({ size = 16, className = '' }) => (
-  <svg
-    width={size}
-    height={size}
-    className={className}
-    viewBox='0 0 15 15'
-    xmlns='http://www.w3.org/2000/svg'
-  >
-    <path
-      d='M11.7816 4.03157C12.0062 3.80702 12.0062 3.44295 11.7816 3.2184C11.5571 2.99385 11.193 2.99385 10.9685 3.2184L7.50005 6.68682L4.03164 3.2184C3.80708 2.99385 3.44301 2.99385 3.21846 3.2184C2.99391 3.44295 2.99391 3.80702 3.21846 4.03157L6.68688 7.49999L3.21846 10.9684C2.99391 11.193 2.99391 11.557 3.21846 11.7816C3.44301 12.0061 3.80708 12.0061 4.03164 11.7816L7.50005 8.31316L10.9685 11.7816C11.193 12.0061 11.5571 12.0061 11.7816 11.7816C12.0062 11.557 12.0062 11.193 11.7816 10.9684L8.31322 7.49999L11.7816 4.03157Z'
-      fill='currentColor'
-      fillRule='evenodd'
-      clipRule='evenodd'
-    />
-  </svg>
-);
