--- app/src/select/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Select';
-export type { SelectOption, SelectProps } from './Select';
