+++ lib/src/components/index.ts
@@ -10,6 +10,7 @@ export * from './pagination';
 export * from './panel';
 export * from './radiogroup';
 export * from './scroll-area';
+export * from './select';
 export * from './separator';
 export * from './skeleton';
 export * from './slider';
