+++ lib/src/components/select/index.ts
@@ -1,0 +1,2 @@
+export { default as Select } from './Select';
+export type { SelectOption, SelectProps } from './Select';
